requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
twilio>=9.0.5
openai>=1.40.0
aiohttp>=3.10.0
python-dotenv>=1.0.1
aiofiles>=23.2.1